+++ src/main/java/com/comuctiva/comuctiva/repositoryes/Tip_DocRepositories.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.comuctiva.comuctiva.models.Tip_Doc;
 
-public interface Tip_DocRepositories extends JpaRepository <Tip_Doc, Short> {
+public interface Tip_DocRepositories extends JpaRepository <Tip_Doc, Integer> {
 
 }
